% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci-package.R
\docType{package}
\name{BioVisReportR-package}
\alias{BioVisReportR}
\alias{BioVisReportR-package}
\title{BioVisReportR: A R SDK for Choppy Report System.}
\description{
A R SDK for Choppy Report System. It contains color, input, config
    module. Color module is used to provide color palettes. Input module is used 
    to handle input data. Config module is used to handle config file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/go-choppy/BioVisReportR}
  \item Report bugs at \url{https://github.com/go-choppy/BioVisReportR/issues}
}

}
\author{
\strong{Maintainer}: Jingcheng Yang \email{yjcyxky@163.com}

}
\keyword{internal}
