% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_basic_theme}
\alias{get_basic_theme}
\title{The Theme Utility}
\usage{
get_basic_theme(size_ratio = 1, angle = 0, ...)
}
\arguments{
\item{size_ratio}{the ratio of the size of title}

\item{angle}{}
}
\description{
Auto generate basic theme for user.
}
\examples{
library("BioVisReportR")
theme <- get_basic_theme(size_ratio = 0.9)
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
