% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_color_func}
\alias{get_color_func}
\title{The Color Utility}
\usage{
get_color_func(palname = "npg", data_class = "discrete", mode = "fill")
}
\arguments{
\item{palname}{palette name
Currently there is 19 available options:
\itemize{
\item \code{"npg"}
\item \code{"aaas"}
\item \code{"nejm"}
\item \code{"lancet"}
\item \code{"jama"}
\item \code{"jco"}
\item \code{"ucscgb"}
\item \code{"d3"}
\item \code{"locuszoom"}
\item \code{"igv"}
\item \code{"cosmic"}
\item \code{"uchicago"}
\item \code{"startrek"}
\item \code{"tron"}
\item \code{"futurama"}
\item \code{"rickandmorty"}
\item \code{"simpsons"}
\item \code{"gsea"}
\item \code{"material"}}
See the \href{https://nanx.me/ggsci/articles/ggsci.html}{Scientific
Journal and Sci-Fi Themed Color Palettes for ggplot2} for details.}

\item{mode}{the corresponding scales
Currently there is 2 available options:
\itemize{
\item \code{"fill"}
\item \code{"color"}}}
}
\description{
Get color function from BioVisReportR namespace.
}
\examples{
library("BioVisReportR")
nejm_fill_func <- get_color_func('nejm', 'fill')
}
\references{
\url{https://nanx.me/ggsci/articles/ggsci.html}
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
