% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_continuous_palette}
\alias{get_continuous_palette}
\title{The Color Utility}
\usage{
get_continuous_palette(palname, palettes = NULL)
}
\arguments{
\item{palname}{the continuous palette name
Currently there is 19 available options:
\itemize{
\item \code{"gsea"}
\item \code{"material"}}
See the \href{https://nanx.me/ggsci/articles/ggsci.html}{Scientific
Journal and Sci-Fi Themed Color Palettes for ggplot2} for details.}

\item{palette}{color palette
Currently there is 20 available options:
\itemize{
\item \code{"blue-grey"}
\item \code{"grey"}
\item \code{"brown"}
\item \code{"deep-orange"}
\item \code{"orange"}
\item \code{"amber"}
\item \code{"yellow"}
\item \code{"lime"}
\item \code{"light-green"}
\item \code{"green"}
\item \code{"teal"}
\item \code{"cyan"}
\item \code{"light-blue"}
\item \code{"blue"}
\item \code{"indigo"}
\item \code{"deep-purple"}
\item \code{"purple"}
\item \code{"red"}
\item \code{"pink"}
\item \code{"default"}}
See the \href{https://nanx.me/ggsci/articles/ggsci.html}{Scientific
Journal and Sci-Fi Themed Color Palettes for ggplot2} for details.}
}
\description{
Show palette supported by Continuous Color Palettes.
}
\examples{
library("BioVisReportR")
palettes <- get_continuous_palette('material', palettes = 'red')
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
