% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_discrete_pal_lst}
\alias{get_discrete_pal_lst}
\title{The Color Utility}
\usage{
get_discrete_pal_lst()
}
\description{
Show all supported palettes.
}
\examples{
library("ChoppyReportR")
palettes <- get_discrete_pal_lst()
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
