% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{get_mode_lst}
\alias{get_mode_lst}
\title{The Color Utility}
\usage{
get_mode_lst()
}
\description{
Show all supported modes
}
\examples{
library("BioVisReportR")
modes <- get_mode_lst()
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
