% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_theme}
\alias{get_theme}
\title{The Theme Utility}
\usage{
get_theme(data, theme_name = "npg", threshold = 10, mode = "fill",
  reverse = False, palette = NULL, alpha = 1)
}
\arguments{
\item{data}{the data that is mapping with color.}

\item{theme_name}{theme name
Currently there is 19 available options:
\itemize{
\item \code{"npg"}
\item \code{"aaas"}
\item \code{"nejm"}
\item \code{"lancet"}
\item \code{"jama"}
\item \code{"jco"}
\item \code{"ucscgb"}
\item \code{"d3"}
\item \code{"locuszoom"}
\item \code{"igv"}
\item \code{"cosmic"}
\item \code{"uchicago"}
\item \code{"startrek"}
\item \code{"tron"}
\item \code{"futurama"}
\item \code{"rickandmorty"}
\item \code{"simpsons"}
\item \code{"gsea"}
\item \code{"material"}}
See the \href{https://nanx.me/ggsci/articles/ggsci.html}{Scientific
Journal and Sci-Fi Themed Color Palettes for ggplot2} for details.}

\item{threshold}{Integer. It is a continuous variable when the number of unique element is greater than threshold}

\item{reverse}{Logical. Should the order of the colors be reversed?}

\item{palette}{Palette type, only when data is a continuous variable.}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\description{
Auto generate theme for user.
}
\examples{
library("ChoppyReportR")
theme <- get_theme(c(1:16), theme_name = 'npg')
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
