% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{is.discrete}
\alias{is.discrete}
\title{The Theme Utility}
\usage{
is.discrete(data, threshold = 10)
}
\arguments{
\item{threshold}{It is a continuous variable when the number of unique element is greater than threshold.}

\item{x}{an object to be tested or coerced.}
}
\description{
Each element is discrete class.
}
\examples{
library("BioVisReportR")
is.discrete(c(1:16))
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
