% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{The Theme Utility}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{an object to be tested or coerced.}
}
\description{
Each element is number class.
}
\examples{
library("BioVisReportR")
is.wholenumber(c(1:16))
}
\author{
Jingcheng Yang <\email{yjcyxky@163.com}>
}
